requirements.txt
flask
flask-cors
python-dotenv
python-jose
requests
httpx==0.27.2